+++ constants.go
@@ -5,7 +5,7 @@ const (
 	AppShortName    = "TBTC"
 	AppProcessName  = "tbtc"
 	AppCreatorName  = "nidea1"
-	AppVersion      = "0.2.0"
+	AppVersion      = "0.3.0"
 	GameProcessName = "TaskBarHero.exe"
 
 	EnablePriceHUD           = true
@@ -27,6 +27,9 @@ const (
 
 	PROCESS_VM_READ           = 0x0010
 	PROCESS_QUERY_INFORMATION = 0x0400
+	SYNCHRONIZE               = 0x00100000
+	INFINITE                  = 0xFFFFFFFF
+	WAIT_OBJECT_0             = 0x00000000
 
 	WS_POPUP          = 0x80000000
 	WS_EX_TOPMOST     = 0x00000008