+++ main.go
@@ -1,5 +1,8 @@
 package main
 
 func main() {
+	if runRestartAfterUpdateHelper() {
+		return
+	}
 	runApp()
 }